Softened cuts to allow two-stage cut process

+++ Skimming/python/MUSiCSkimmer_cfi.py
@@ -60,10 +60,10 @@ Skimmer = cms.EDAnalyzer(
 
     cuts = cms.PSet(
         min_muon_pt = cms.double( 5 ),
-        min_ele_pt = cms.double( 15 ),
-        min_gamma_pt = cms.double( 15 ),
-        min_jet_pt = cms.double( 80 ),
-        min_met = cms.double( 45 ),
+        min_ele_pt = cms.double( 5 ),
+        min_gamma_pt = cms.double( 5 ),
+        min_jet_pt = cms.double( 20 ),
+        min_met = cms.double( 20 ),
         max_eta = cms.double( 3 ),
         min_rechit_energy = cms.double( 20 ),
         min_rechit_swiss_cross = cms.double( 0.8 ),
